// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CustomerPlan = {
  free: 'free',
  pro: 'pro',
  enterprise: 'enterprise'
} as const

export type CustomerPlan = (typeof CustomerPlan)[keyof typeof CustomerPlan]


export const TicketStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const TicketCategory = {
  billing: 'billing',
  bug: 'bug',
  account_access: 'account_access',
  feature_request: 'feature_request',
  other: 'other'
} as const

export type TicketCategory = (typeof TicketCategory)[keyof typeof TicketCategory]


export const TicketPriority = {
  P0: 'P0',
  P1: 'P1',
  P2: 'P2',
  P3: 'P3'
} as const

export type TicketPriority = (typeof TicketPriority)[keyof typeof TicketPriority]


export const TriageDecision = {
  auto_accept: 'auto_accept',
  manual_review: 'manual_review'
} as const

export type TriageDecision = (typeof TriageDecision)[keyof typeof TriageDecision]